wip: user interface rehaul part 7

+++ libs/feat/execution-settings/src/lib/execution-settings.ts
@@ -2,27 +2,28 @@ import { ApplicationLibrary, Clerk } from '@worksheets/apps/framework';
 import { newApplicationsDatabase } from '@worksheets/data-access/applications';
 import { newHandshakesDatabase } from '@worksheets/data-access/handshakes';
 import {
-  SettingEntity,
-  newSettingsDatabase,
+  ConnectionEntity,
+  newConnectionDatabase,
 } from '@worksheets/data-access/settings';
 import { HandlerFailure } from '@worksheets/util/next';
-import { OAuthClient, OAuthToken } from '@worksheets/util/oauth/client';
-import {
-  closeRedirect,
-  errorRedirect,
-  findOAuthProperty,
-  findProperty,
-} from './util';
+import { OAuthClient } from '@worksheets/util/oauth/client';
+import { closeRedirect, errorRedirect } from './util';
 import { isExpired } from '@worksheets/util/time';
+import { TRPCError } from '@trpc/server';
+import { RequiredBy } from '@worksheets/util/types';
+import { z } from 'zod';
 
-const settingsDb = newSettingsDatabase();
-const applicationsDb = newApplicationsDatabase();
+const connectionsDb = newConnectionDatabase();
+const registry = newApplicationsDatabase();
 const handshakesDb = newHandshakesDatabase();
 
 export const newPrivateLibrary = (userId: string) => {
   const library = new ApplicationLibrary({
-    clerk: applicationsDb,
-    settingsLoader: (methodPath) => getSettings(userId, methodPath),
+    clerk: registry,
+    settingsLoader: (methodPath) => {
+      console.info('TODO', userId, methodPath);
+      return Promise.resolve({});
+    },
   });
   return library;
 };
@@ -34,124 +35,146 @@ export const newEmptyLibrary = () => {
   });
 };
 
-export const findSettingEntityId = async (
-  userId: string,
-  methodPath: string,
-  propertyKey: string
-): Promise<string | undefined> => {
-  const settings = await settingsDb.query(
-    { f: 'uid', o: '==', v: userId },
-    { f: 'method', o: '==', v: methodPath },
-    { f: 'key', o: '==', v: propertyKey }
-  );
+// const getLatestTokens = async (setting: SettingEntity): Promise<OAuthToken> => {
+//   if (setting.type !== 'oauth') {
+//     throw new HandlerFailure({
+//       code: 'unsupported-operation',
+//       message: `invalid setting type, expected 'oauth' received '${setting.type}' `,
+//     });
+//   }
 
-  if (!settings || settings.length < 1) {
-    return undefined;
+//   if (!setting.data) {
+//     throw new HandlerFailure({
+//       code: 'unsupported-operation',
+//       message: `setting must have a token set`,
+//     });
+//   }
+
+//   const method = applicationsDb.borrow(setting.method);
+
+//   const prop = findOAuthProperty(method, setting.key);
+
+//   const client = new OAuthClient(prop.options);
+
+//   const tokens = client.convertToOAuthToken(setting.data as string);
+
+//   if (tokens.expired() && tokens.refreshToken) {
+//     const refreshed = await tokens.refresh();
+
+//     setting.data = client.serializeToken(refreshed);
+
+//     await settingsDb.update({ ...setting });
+
+//     return refreshed;
+//   }
+
+//   return tokens;
+// };
+
+export const connectionFormSchema = z.object({
+  id: z.string().describe('the connection id'),
+  name: z.string().optional(),
+  appId: z.string(),
+  settings: z.record(z.unknown()).optional(),
+});
+
+export type ConnectionForm = z.infer<typeof connectionFormSchema>;
+
+export const loadConnectionForm = async ({
+  id,
+}: {
+  id: string;
+  uid: string;
+}): Promise<ConnectionForm> => {
+  // return empty form if connection doesn't exist
+  if (!id || !(await connectionsDb.has(id))) {
+    return {
+      id: '',
+      name: '',
+      appId: '',
+      settings: {},
+    };
   }
+  const connection = await connectionsDb.get(id);
+  const app = registry.getApp(connection.appId);
+  // get settings
+  const settings = app.settings;
+  // for each connection setting, replace oauth tokens if they are set.
+  if (settings && connection.settings) {
+    connection.settings = Object.keys(connection.settings).reduce(
+      (acc, key) => {
+        const value = connection.settings[key];
+        const setting = settings[key];
+        if (setting.type === 'oauth') {
+          acc[key] = Boolean(value);
+        } else {
+          acc[key] = value;
+        }
 
-  if (settings.length > 1) {
-    console.warn(
-      `unexpected: found more than 1 settings with the same method, uid, and key`,
-      settings.map((s) => s.id)
-    );
-  }
-
-  return settings[0].id;
-};
-
-export const upsertSetting = async (
-  userId: string,
-  methodPath: string,
-  propertyKey: string,
-  data?: unknown
-): Promise<{ url?: string }> => {
-  const method = applicationsDb.borrow(methodPath);
-  const prop = findProperty(method, propertyKey);
-  const entityId = await findSettingEntityId(userId, methodPath, propertyKey);
-  if (prop.type === 'flag' || prop.type === 'token') {
-    await settingsDb.updateOrInsert({
-      id: entityId ?? settingsDb.id(),
-      uid: userId,
-      type: prop.type,
-      method: methodPath,
-      key: propertyKey,
-      data: data,
-    });
-
-    return {};
-  }
-
-  if (prop.type === 'oauth') {
-    const client = new OAuthClient(prop.options);
-
-    const handshakeId = await createHandshake(
-      userId,
-      entityId,
-      methodPath,
-      propertyKey
-    );
-
-    const url = client.getUri(handshakeId);
-
-    return { url };
-  }
-
-  throw new HandlerFailure({
-    code: 'unsupported-operation',
-    message: `cannot save property type`,
-    data: prop,
-  });
-};
-
-export const saveSetting = async (opt: Omit<SettingEntity, 'id'>) => {
-  const id = settingsDb.id();
-  await settingsDb.insert({
-    ...opt,
-    id,
-  });
-};
-
-export const getSettings = async (
-  uid: string,
-  methodPath: string
-): Promise<Record<string, unknown>> => {
-  if (!uid) {
-    throw new HandlerFailure({
-      code: 'unauthorized',
-      message: 'cannot get settings without a user associated',
-      data: {
-        uid,
-        methodPath,
-        possibleReasons:
-          'caused by an anonymous worksheet execution that tried to accesss a method that requires global account settings',
+        return acc;
       },
+      {} as Record<string, unknown>
+    );
+  }
+
+  return {
+    id: connection.id,
+    name: connection.name,
+    appId: connection.appId,
+    settings: connection.settings,
+  };
+};
+
+export const listConnections = async ({ uid }: { uid: string }) => {
+  const connections = await connectionsDb.query({ f: 'uid', o: '==', v: uid });
+  return connections;
+};
+
+export const deleteConnectionField = async ({
+  connectionId,
+  settingId,
+}: {
+  settingId: string;
+  connectionId: string;
+}) => {
+  const connection = await connectionsDb.get(connectionId);
+
+  if (!connection) {
+    throw new TRPCError({
+      code: 'BAD_REQUEST',
+      message: `invalid connection id '${connectionId}'`,
     });
   }
 
-  const settings = await settingsDb.query(
-    { f: 'uid', o: '==', v: uid },
-    { f: 'method', o: '==', v: methodPath }
-  );
+  // find the setting on the connection.
+  console.log('deleting setting', settingId);
+  connection.settings[settingId] = '';
 
-  console.info(
-    `got ${settings.length} settings for method ${methodPath} for user ${uid}`
-  );
+  await applyConnectionUpdates(connectionId, {
+    settings: connection.settings,
+  });
 
-  if (settings.length === 0) {
-    return {};
-  }
-
-  const output: Record<string, unknown> = {};
-  for (const setting of settings) {
-    const data =
-      setting.type !== 'oauth' ? setting.data : await getLatestTokens(setting);
-
-    output[setting.key] = data;
-  }
-  return output;
+  return connection;
 };
-
-export const saveOAuthSetting = async (
+export const deleteConnection = async ({
+  id,
+  uid,
+}: {
+  id: string;
+  uid: string;
+}) => {
+  const connection = await connectionsDb.get(id);
+  //check if user has access
+  if (connection.uid !== uid) {
+    throw new TRPCError({
+      code: 'UNAUTHORIZED',
+      message: `user does not have access to this connection`,
+    });
+  }
+  await connectionsDb.delete(id);
+  return true;
+};
+export const resolveHandshake = async (
   url?: string,
   handshakeId?: string
 ): Promise<{ url: string }> => {
@@ -168,26 +191,30 @@ export const saveOAuthSetting = async (
     return errorRedirect('INVALID_HANDSHAKE');
   }
 
-  const { methodPath, propertyKey, uid, timestamp, settingId } = handshake;
+  const { appId, uid, timestamp, settingId, connectionId } = handshake;
   if (isExpired(timestamp + 10 * 60 * 1000)) {
     return errorRedirect('EXPIRED_HANDSHAKE');
   }
 
-  const method = applicationsDb.borrow(methodPath);
-  const prop = findOAuthProperty(method, propertyKey);
+  const app = registry.getApp(appId);
+  const setting = app?.settings?.[settingId];
+
+  if (!setting) {
+    return errorRedirect('INVALID_SETTING');
+  }
+
+  if (setting.type !== 'oauth') {
+    return errorRedirect('INVALID_SETTING_TYPE');
+  }
 
   try {
-    const client = new OAuthClient(prop.options);
+    const client = new OAuthClient(setting.options);
     const tokens = await client.parseUrl(url);
 
-    const id = settingId ?? settingsDb.id();
-    await settingsDb.updateOrInsert({
-      id,
+    await applyConnectionUpdates(connectionId, {
       uid: uid,
-      type: 'oauth',
-      method: methodPath,
-      key: propertyKey,
-      data: tokens,
+      appId: appId,
+      settings: { [settingId]: tokens },
     });
 
     await deleteHandshake(handshakeId);
@@ -201,57 +228,104 @@ export const saveOAuthSetting = async (
   }
 };
 
-const getLatestTokens = async (setting: SettingEntity): Promise<OAuthToken> => {
+export const createOAuthUrl = async (opts: ConnectionFieldOptions) => {
+  const { appId, settingId } = opts;
+
+  const app = registry.getApp(appId);
+
+  const settings = app.settings;
+  if (!settings) {
+    throw new TRPCError({
+      code: 'BAD_REQUEST',
+      message: `cannot create oauth url for app ${appId} because it does not have any settings`,
+    });
+  }
+
+  const setting = settings[settingId];
   if (setting.type !== 'oauth') {
-    throw new HandlerFailure({
-      code: 'unsupported-operation',
-      message: `invalid setting type, expected 'oauth' received '${setting.type}' `,
+    throw new TRPCError({
+      code: 'BAD_REQUEST',
+      message: 'setting is not of type oauth',
     });
   }
 
-  if (!setting.data) {
-    throw new HandlerFailure({
-      code: 'unsupported-operation',
-      message: `setting must have a token set`,
-    });
-  }
+  const client = new OAuthClient(setting.options);
 
-  const method = applicationsDb.borrow(setting.method);
+  const handshakeId = await createHandshake(opts);
 
-  const prop = findOAuthProperty(method, setting.key);
+  const url = client.getUri(handshakeId);
 
-  const client = new OAuthClient(prop.options);
-
-  const tokens = client.convertToOAuthToken(setting.data as string);
-
-  if (tokens.expired() && tokens.refreshToken) {
-    const refreshed = await tokens.refresh();
-
-    setting.data = client.serializeToken(refreshed);
-
-    await settingsDb.update({ ...setting });
-
-    return refreshed;
-  }
-
-  return tokens;
+  return { url };
 };
 
-const createHandshake = async (
-  userId: string,
-  settingId: string | undefined,
-  methodPath: string,
-  propertyKey: string
+export const submitConnectionForm = async (
+  entity: RequiredBy<ConnectionEntity, 'id' | 'appId'>
 ) => {
+  const settings = entity.settings;
+  if (settings) {
+    // get the app from the registry
+    const app = registry.getApp(entity.appId);
+
+    // get the settings from the app
+    const fields = app.settings;
+    if (!fields) {
+      throw new TRPCError({
+        code: 'BAD_REQUEST',
+        message: `cannot save connection form for app ${entity.appId} because it does not have any settings`,
+      });
+    }
+
+    // remove any oauth settings from the entity
+    const obj = Object.keys(settings)
+      .filter((key) => fields[key]?.type !== 'oauth')
+      .reduce((acc, key) => {
+        acc[key] = settings[key];
+        return acc;
+      }, {} as Record<string, unknown>);
+
+    entity.settings = obj;
+  }
+
+  await applyConnectionUpdates(entity.id, entity);
+};
+
+const applyConnectionUpdates = async (
+  connectionId: string,
+  { uid, appId, settings, name }: Partial<Omit<ConnectionEntity, 'id'>>
+) => {
+  await connectionsDb.apply(connectionId, {
+    uid,
+    appId,
+    name,
+    settings,
+    updatedAt: Date.now(),
+  });
+};
+
+type ConnectionFieldOptions = {
+  userId: string;
+  connectionId: string;
+  appId: string;
+  settingId: string;
+};
+
+const createHandshake = async ({
+  userId,
+  connectionId,
+  appId,
+  settingId,
+}: ConnectionFieldOptions) => {
   const handshakeId = handshakesDb.id();
+
   await handshakesDb.insert({
     uid: userId,
     id: handshakeId,
-    settingId: settingId,
     timestamp: Date.now(),
-    methodPath,
-    propertyKey,
+    connectionId,
+    appId,
+    settingId,
   });
+
   return handshakeId;
 };
